requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
cryptography==41.0.4
Flask-JWT-Extended==4.5.3
python-ulid==1.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.0.0
bcrypt==4.0.1
requests==2.31.0
gunicorn==21.2.0